Initialize the Find Data dialog with the selection.
This should resolve bug 10962.

Whenever the Find Data dialog is shown (via the Edit->Find Data menu
item of the Find by Scalar Value toolbar button), the current selection
is used as the source object; the active pipeline output port is only
used when no selection exists.
Furthermore, if the current selection is a vtkQuerySelectionSource, then
the query is parsed and the selection type and clauses are populated
to match the previous query. Note that only queries generated by the
Find Data dialog's UI can be parsed.

Change-Id: Idf9b14336956c748bcd8954a7400ace094405600